Added namespacing and a single data type settings header.
Added proper namespacing. For the classes, this namespacing is
consistent with the file structure, which also changed slightly (RLWE is
now Main_Classes). The global namespace for this project is
RLWE_Toolkit. All functions that were global are now in this namespace.

Further I added a singel header that defines the used data types for the
integer and real numbers and some useful macros.